add argparse

+++ script/mdict_es.py
@@ -4,6 +4,7 @@ import psutil
 import time
 import hashlib
 import zlib
+import argparse
 
 current_path = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(os.path.dirname(current_path))
@@ -17,16 +18,39 @@ from mdict.mdict_utils.search_object import SearchObject
 from mdict.models import MdictDic
 from mdict.mdict_utils.init_utils import init_vars
 
+from elasticsearch import Elasticsearch
 from elasticsearch.helpers import parallel_bulk
 from elasticsearch.helpers.errors import BulkIndexError
 from elasticsearch_dsl import connections, analyzer, Text, Index, Document
-from elasticsearch.exceptions import TransportError, NotFoundError
+from elasticsearch.exceptions import TransportError, ConnectionError, NotFoundError
+
+parser = argparse.ArgumentParser()
+par = parser.add_mutually_exclusive_group()
+par.add_argument('-c', '--create', nargs='+', default=[], help='create mdict index with id list')
+par.add_argument('-ca', '--createall', action='store_false', help='create all mdict index')
+par.add_argument('-d', '--delete', nargs='+', default=[], help='delete mdict index with id list')
+par.add_argument('-da', '--deleteall', action='store_true', help='delete all mdict index')
+args = parser.parse_args()
+create_param = args.create
+createall_param = args.createall
+delete_param = args.delete
+deleteall_param = args.deleteall
+
+if not createall_param:
+    createall_param = True
+
+if deleteall_param:
+    createall_param = False
+
+if create_param:
+    createall_param = False
+
+if delete_param:
+    createall_param = False
 
 error_log_path = os.path.join(current_path, 'error.log')
 
 cpu_num = psutil.cpu_count(False)
-
-connections.create_connection()
 
 html_strip1 = analyzer('html_strip',
                        tokenizer="ik_smart",
@@ -80,15 +104,23 @@ def get_md5(mdx):
 
 
 def get_md5_with_pk(dic_pk):
-    dic = MdictDic.objects.get(pk=dic_pk)
-    md5 = dic.mdict_md5
-    if md5 == '':
-        item = init_vars.mdict_odict[dic.mdict_file]
-        mdx = item.mdx
-        md5 = get_md5(mdx)
-        dic.mdict_md5 = md5
-        dic.save()
-    return md5
+    dics = MdictDic.objects.filter(pk=dic_pk)
+    if len(dics) > 0:
+        dic = dics[0]
+        md5 = dic.mdict_md5
+        od = init_vars.mdict_odict
+        if md5 == '':
+            if dic.mdict_file in od.keys():
+                item = od[dic.mdict_file]
+                mdx = item.mdx
+                md5 = get_md5(mdx)
+                dic.mdict_md5 = md5
+                dic.save()
+            else:
+                return ''
+        return md5
+    else:
+        return ''
 
 
 def create_index(dic, md5):
@@ -236,14 +268,52 @@ def create_cache(dic, mdx):
         index.put_settings(body=settings2)
 
 
-def close_index(dic_pk):
-    index = Index(get_index_name_with_pk(dic_pk))
-    index.close()
+def close_index_with_pk(dic_pk):
+    index_name = get_index_name_with_pk(dic_pk)
+    if index_name == 'mdict-':
+        print(dic_pk, 'not exists')
+        return
+    index = Index(index_name)
+    try:
+        index.close()
+    except NotFoundError as e:
+        print(e)
+    print('close', dic_pk, index_name)
 
 
-def delete_index(dic_pk):
-    index = Index(get_index_name_with_pk(dic_pk))
-    index.delete()
+
+def open_index_with_pk(dic_pk):
+    index_name = get_index_name_with_pk(dic_pk)
+    if index_name == 'mdict-':
+        print(dic_pk, 'not exists')
+        return
+    index = Index(index_name)
+    try:
+        index.open()
+    except NotFoundError as e:
+        print(e)
+    print('open', dic_pk, index_name)
+
+
+def delete_index_with_pk(dic_pk):
+    index_name = get_index_name_with_pk(dic_pk)
+    if index_name == 'mdict-':
+        print(dic_pk, 'not exists')
+        return
+    index = Index(index_name)
+    try:
+        index.delete()
+    except NotFoundError as e:
+        print(e)
+    print('delete', dic_pk, index_name)
+
+
+def delete_all_es():
+    client = Elasticsearch()
+    for index in client.indices.get_alias('mdict-*'):
+        index = Index(index)
+        index.delete()
+        print('delete', index)
 
 
 def create_es_with_pk(dic_pk):
@@ -276,7 +346,7 @@ def create_all_es(pk_list=[]):
     t0 = time.perf_counter()
     odict = init_vars.mdict_odict
     odict_len = len(odict)
-    i = 0
+    i = 1
     for k in odict.keys():
         item = odict[k]
         mdx = item.mdx
@@ -322,8 +392,30 @@ def create_all_es(pk_list=[]):
     print('indexing time', t3 - t0)
 
 
-# create_es_with_pk(934)
+if __name__ == '__main__':
+    print('create_param', create_param)
+    print('createall_param', createall_param)
+    print('delete_param', delete_param)
+    print('deleteall_param', deleteall_param)
+    try:
+        connections.create_connection()
+        if delete_param:
+            delete_list = []
+            for d in delete_param:
+                delete_list.append(int(d))
+            for d in delete_list:
+                delete_index_with_pk(d)
+        elif create_param:
+            create_list = []
+            for c in create_param:
+                create_list.append(int(c))
+            for c in create_list:
+                create_es_with_pk(c)
+            # create_all_es(pk_list=create_list)
+        elif createall_param:
+            create_all_es()
+        elif deleteall_param:
+            delete_all_es()
 
-# delete_index(743)
-
-create_all_es()
+    except ConnectionError as e:
+        print(e)
